Moved colors to a file
Just a code cleanup
