+++ scripts/mergers/mergers.py
@@ -27,6 +27,7 @@ import sys
 from multiprocessing import cpu_count
 from threading import Lock
 from concurrent.futures import ThreadPoolExecutor, as_completed
+from scripts.mergers.bcolors import bcolors
 
 from inspect import currentframe
 
@@ -38,17 +39,6 @@ def tryit(func):
         func() 
     except:
         pass 
-
-class bcolors:
-    HEADER = '\033[95m'
-    OKBLUE = '\033[94m'
-    OKCYAN = '\033[96m'
-    OKGREEN = '\033[92m'
-    WARNING = '\033[93m'
-    FAIL = '\033[91m'
-    ENDC = '\033[0m'
-    BOLD = '\033[1m'
-    UNDERLINE = '\033[4m'
 
 stopmerge = False
